# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Sales BI Platform — unified API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /datasets:
    get:
      operationId: listDatasets
      tags: [datasets]
      summary: List datasets for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dataset"
    post:
      operationId: createDataset
      tags: [datasets]
      summary: Create dataset record (after upload)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatasetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /datasets/{datasetId}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get dataset by ID
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
    delete:
      operationId: deleteDataset
      tags: [datasets]
      summary: Delete dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /datasets/{datasetId}/summary:
    get:
      operationId: getDatasetSummary
      tags: [datasets]
      summary: Get dataset summary stats
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetSummary"
  /datasets/upload:
    post:
      operationId: uploadDataset
      tags: [datasets]
      summary: Upload Excel/CSV file and parse into DB
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DatasetFileUpload"
      responses:
        "201":
          description: Uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"

  # ─── KPI ─────────────────────────────────────────────────────────
...
  /dashboards:
    get:
      operationId: listDashboards
      tags: [dashboards]
      summary: List saved dashboards for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dashboard"
    post:
      operationId: createDashboard
      tags: [dashboards]
      summary: Save a new dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /dashboards/{dashboardId}:
    get:
      operationId: getDashboard
      tags: [dashboards]
      summary: Get dashboard by ID
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
    patch:
      operationId: updateDashboard
      tags: [dashboards]
      summary: Update dashboard
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
    delete:
      operationId: deleteDashboard
      tags: [dashboards]
      summary: Delete dashboard
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Map ─────────────────────────────────────────────────────────
...
  /export/excel:
    post:
      operationId: exportExcel
      tags: [export]
      summary: Export KPI data as Excel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /export/csv:
    post:
      operationId: exportCsv
      tags: [export]
      summary: Export KPI data as CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /export/json:
    post:
      operationId: exportJson
      tags: [export]
      summary: Export KPI data as JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: JSON file
          content:
            application/json:
              schema:
                type: object

  # ─── Agents ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Dataset:
      type: object
      required: [id, name, userId, rowCount, columns, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        userId:
          type: string
        fileName:
          type: string
        rowCount:
          type: integer
        columns:
          type: array
          items:
            type: string
        status:
          type: string
        isDemo:
          type: boolean
        createdAt:
          type: string
    DatasetInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        fileName:
          type: string
        rowCount:
          type: integer
        columns:
          type: array
          items:
            type: string
    DatasetSummary:
      type: object
      required: [datasetId, totalRevenue, totalGrossProfit, totalEbitda, totalRoi, countRows, countries, years]
      properties:
        datasetId:
          type: string
        totalRevenue:
          type: number
        totalGrossProfit:
          type: number
        totalEbitda:
          type: number
        totalRoi:
          type: number
        countRows:
          type: integer
        countries:
          type: array
          items:
            type: string
        years:
          type: array
          items:
            type: integer
        categories:
          type: array
          items:
            type: string

    # ─── KPI ───────────────────────────────────────────────────────
    KpiFilters:
      type: object
      properties:
        country:
          type: ["string", "null"]
        branch:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        month:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
...
    Dashboard:
      type: object
      required: [id, userId, name, config, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        config:
          type: object
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    DashboardInput:
      type: object
      required: [name, config]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        config:
          type: object
        isPublic:
          type: boolean
    DashboardUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        config:
          type: object
        isPublic:
          type: boolean

    # ─── Map ───────────────────────────────────────────────────────
...
    ExportInput:
      type: object
      required: [datasetId, type]
      properties:
        datasetId:
          type: string
        type:
          type: string
          description: "kpi | pl | bcg | forecast | raw"
        filters:
          $ref: "#/components/schemas/KpiFilters"
        groupBy:
          type: array
          items:
            type: string

    # ─── Agents ────────────────────────────────────────────────────
...
    DatasetFileUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        name:
          type: string
        sheet:
          type: string
...